requirements.txt
# eink
spidev
RPi.GPIO
qrcode
Pillow

# web
Flask
